new tests and improved internal consitency
