+++ astropy/coordinates/tests/test_intermediate_transformations.py
@@ -17,7 +17,7 @@ from astropy.coordinates import (
     HCRS, HeliocentricMeanEcliptic, TEME, TETE)
 from astropy.coordinates.solar_system import _apparent_position_in_true_coordinates
 from astropy.utils import iers
-from astropy.utils.exceptions import AstropyWarning
+from astropy.utils.exceptions import AstropyWarning, AstropyDeprecationWarning
 
 from .utils import randomly_sample_sphere
 from astropy.coordinates.builtin_frames.utils import get_jd12
@@ -612,21 +612,23 @@ def test_tete_transforms():
     # test self-transform by comparing to GCRS-TETE-ITRS-TETE route
     tete_coo1 = moon.transform_to(tete_frame)
     tete_coo2 = moon.transform_to(tete_geo)
-    assert_allclose(tete_coo1.separation_3d(tete_coo2), 0*u.mm, atol=100*u.mm)
+    assert_allclose(tete_coo1.separation_3d(tete_coo2), 0*u.mm, atol=1*u.mm)
 
     # test TETE-ITRS transform by comparing GCRS-CIRS-ITRS to GCRS-TETE-ITRS
     itrs1 = moon.transform_to(CIRS()).transform_to(ITRS())
     itrs2 = moon.transform_to(TETE()).transform_to(ITRS())
-    assert_allclose(itrs1.separation_3d(itrs2), 0*u.mm, atol=100*u.mm)
+    assert_allclose(itrs1.separation_3d(itrs2), 0*u.mm, atol=1*u.mm)
 
     # test round trip GCRS->TETE->GCRS
     new_moon = moon.transform_to(TETE()).transform_to(moon)
-    assert_allclose(new_moon.separation_3d(moon), 0*u.mm, atol=100*u.mm)
+    assert_allclose(new_moon.separation_3d(moon), 0*u.mm, atol=1*u.mm)
 
     # test round trip via ITRS
     tete_rt = tete_coo1.transform_to(ITRS(obstime=time)).transform_to(tete_coo1)
-    assert_allclose(tete_rt.separation_3d(tete_coo1), 0*u.mm, atol=100*u.mm)
+    assert_allclose(tete_rt.separation_3d(tete_coo1), 0*u.mm, atol=1*u.mm)
 
     # ensure deprecated routine remains consistent
-    tete_alt = _apparent_position_in_true_coordinates(moon)
+    # make sure test raises warning!
+    with pytest.warns(AstropyDeprecationWarning, match='The use of'):
+        tete_alt = _apparent_position_in_true_coordinates(moon)
     assert_allclose(tete_coo1.separation_3d(tete_alt), 0*u.mm, atol=100*u.mm)
